+++ package.json
@@ -7,11 +7,12 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.4",
-    "bootstrap": "^5.0.2",
+    "bootstrap": "^5.2.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "firebase": "^8.7.0",
+    "firebase": "^8.10.1",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.5.0",
     "react-dom": "^17.0.2",
     "react-icons-kit": "^1.3.1",
     "react-router-dom": "^5.2.0",
