--- src/Components/DriverDashboard.js
@@ -1,39 +1,0 @@
-import React,{ useState} from 'react'
-import { Card, Button, Alert } from 'react-bootstrap'
-import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
-
-
-export default function Dashboard() {
-    const [error, setError] = useState("")
-    const { currentUser, logout } = useAuth()
-    const navigate = useNavigate()
-
-
-    async function handleLogout(){
-      setError('')
-
-      try {
-        await logout()
-        navigate('/driverlogin')
-      } catch {
-        setError('Failed to log out')
-      }
-    }
-
-  return (
-    <>
-      <Card>
-        <Card.Body>
-          <h2 className='text-center mb-4'>Profile</h2>
-          {error && <Alert variant='danger'>{error}</Alert>}
-          <strong>Driver:</strong> {currentUser.email}
-          <Link to='/update-profile' className='btn btn-primary w-100 mt-3'>Update Profile</Link>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        <Button variant='link' to='/driverlogin' onClick={handleLogout}>Log Out</Button>
-      </div>
-    </>
-  )
-}
