--- src/Components/DriverSignup.js
@@ -1,82 +1,0 @@
-//import { Alert } from 'bootstrap'
-import React, { useRef, useState } from 'react'
-import { Form, Button, Card, Alert } from 'react-bootstrap'
-import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
-
-
-
-export default function Signup() {
-    const emailRef = useRef()
-    const passwordRef = useRef()
-    const passwordConfirmRef = useRef()
-    const nameRef = useRef()
-    const surnameRef = useRef()
-    const drivernumberRef = useRef()
-    const { signup } = useAuth()
-    const [error, setError] = useState('')
-    const [loading, setLoading] = useState(false)
-    const navigate = useNavigate()
-
-    async function handleSubmit(e) {
-        e.preventDefault()
-
-        if (passwordRef.current.value !== passwordConfirmRef.current.value) {
-            return setError('Passwords do not match')    
-        }
-
-        try{
-            setError('')
-            setLoading(true)
-            await signup(emailRef.current.value, passwordRef.current.value)
-            navigate('/')
-        } catch {
-            setError('Failed to create an account. Password length less than 8 characters.')
-        }
-        setLoading(false)
-       
-    }
-
-  return (
-    <>
-      <Card>
-        <Card.Body>
-            <h2 className='text-center mb-4'>Sign Up</h2>
-           
-            {error && <Alert variant='danger'>{error}</Alert>}
-            <Form onSubmit={handleSubmit}>
-              <Form.Group id='name'>
-                    <Form.Label>Name</Form.Label>
-                    <Form.Control type='text' ref={nameRef} required/>
-                </Form.Group>
-                <Form.Group id='surname'>
-                    <Form.Label>Surname</Form.Label>
-                    <Form.Control type='text' ref={surnameRef} required/>
-                </Form.Group>
-                <Form.Group id='drivernumber'>
-                    <Form.Label>Driver License Number</Form.Label>
-                    <Form.Control type='number' ref={drivernumberRef} required/>
-                </Form.Group>
-                <Form.Group id='email'>
-                    <Form.Label>Email</Form.Label>
-                    <Form.Control type='email' ref={emailRef} required/>
-                </Form.Group>
-                <Form.Group id='password'>
-                    <Form.Label>Password</Form.Label>
-                    <Form.Control type='password' ref={passwordRef} required/>
-                </Form.Group>
-                <Form.Group id='password-confirm'>
-                    <Form.Label>Password Confirmation</Form.Label>
-                    <Form.Control type='password' ref={passwordConfirmRef} required/>
-                </Form.Group>
-                <Button disabled={loading} className='w-100' type='submit'>Sign Up
-                </Button>
-            </Form>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        Already have an account? <Link to='/driverlogin'>Login In</Link>
-      </div>
-    </>
-  )
-}
