+++ src/Components/ForgotPassword.js
@@ -1,10 +1,10 @@
 import React, { useRef, useState } from 'react'
 import { Form, Button, Card, Alert } from 'react-bootstrap'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from './AuthContext'
 import { Link } from 'react-router-dom'
 
 
-export default function ForgotPassword() {
+export function ForgotPassword() {
     const emailRef = useRef()
     const { resetPassword } = useAuth()
     const [error, setError] = useState('')
