+++ src/Components/Login.js
@@ -1,65 +1,64 @@
-import React, { useRef, useState } from 'react'
-import { Form, Button, Card, Alert } from 'react-bootstrap'
-import { useAuth } from '../contexts/AuthContext'
-import { Link, useNavigate } from 'react-router-dom'
+import React,{useState} from 'react'
+import {Link} from 'react-router-dom'
+import {auth} from '../Config/Config'
+import {useHistory} from 'react-router-dom'
 
+export const Login = () => {
 
-export default function Login() {
-    const emailRef = useRef()
-    const passwordRef = useRef()
-    const { login } = useAuth()
-    const [error, setError] = useState('')
-    const [loading, setLoading] = useState(false)
-    const navigate = useNavigate()
+    const history = useHistory();
 
-    async function handleSubmit(e) {
-        e.preventDefault()
+    const [email, setEmail]=useState('');
+    const [password, setPassword]=useState('');
 
-    
-        try{
-            setError('')
-            setLoading(true)
-            await login(emailRef.current.value, passwordRef.current.value)
-            navigate('/')
-        } catch {
-            setError('Failed to login')
-        }
-        setLoading(false)
-       
+    const [errorMsg, setErrorMsg]=useState('');
+    const [successMsg, setSuccessMsg]=useState('');
+
+    const handleLogin=(e)=>{
+        e.preventDefault();
+        // console.log(email, password);
+        auth.signInWithEmailAndPassword(email,password).then(()=>{
+            setSuccessMsg('Login Successfull. You will now automatically get redirected to Home page');
+            setEmail('');
+            setPassword('');
+            setErrorMsg('');
+            setTimeout(()=>{
+                setSuccessMsg('');
+                history.push('/');
+            },3000)
+        }).catch(error=>setErrorMsg(error.message));
     }
 
-  return (
-    <>
-      <Card>
-        <Card.Body>
-            <h2 className='text-center mb-4'>Log In</h2>
-           
-            {error && <Alert variant='danger'>{error}</Alert>}
-            <Form onSubmit={handleSubmit}>
-                <Form.Group id='email'>
-                    <Form.Label>Email</Form.Label>
-                    <Form.Control type='email' ref={emailRef} required/>
-                </Form.Group>
-                <Form.Group id='password'>
-                    <Form.Label>Password</Form.Label>
-                    <Form.Control type='password' ref={passwordRef} required/>
-                </Form.Group>
-                
-                <Button disabled={loading} className='w-100' type='submit'>Log In
-                </Button>
-            </Form>
-            <div className='w-100 text-center mt-2'>
-                <Link to='/forgot-password'> Forgot Password?</Link>
-            </div>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        Need an account? <Link to='/signup'>Sign Up</Link>
-      </div>
-      <div className='w-100 text-center mx-10 '>
-        Are you a driver? <Link to='/driverlogin'>Login In</Link>
-      </div>
-    </>
-  )
-}
+    return (
+        <div className='container'>
+            <br></br>
+            <br></br>
+            <h1>Login</h1>
+            <hr></hr>
+            {successMsg&&<>
+                <div className='success-msg'>{successMsg}</div>
+                <br></br>
+            </>}
+            <form className='form-group' autoComplete="off"
+            onSubmit={handleLogin}>               
+                <label>Email</label>
+                <input type="email" className='form-control' required
+                onChange={(e)=>setEmail(e.target.value)} value={email}></input>
+                <br></br>
+                <label>Password</label>
+                <input type="password" className='form-control' required
+                onChange={(e)=>setPassword(e.target.value)} value={password}></input>
+                <br></br>
+                <div className='btn-box'>
+                    <span>Don't have an account SignUp
+                    <Link to="signup" className='link'> Here</Link></span>
 
+                    <button type="submit" className='btn btn-success btn-md'>LOGIN</button>
+                </div>
+            </form>
+            {errorMsg&&<>
+                <br></br>
+                <div className='error-msg'>{errorMsg}</div>                
+            </>}
+        </div>
+    )
+}
