+++ src/Components/Signup.js
@@ -1,81 +1,77 @@
-//import { Alert } from 'bootstrap'
-import React, { useRef, useState } from 'react'
-import { Form, Button, Card, Alert } from 'react-bootstrap'
-import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
+import React,{useState} from 'react'
+import {auth,fs} from '../Config/Config'
+import {Link} from 'react-router-dom'
+import {useHistory} from 'react-router-dom'
 
+export const Signup = () => {
 
+    const history = useHistory();
 
-export default function Signup() {
-    const emailRef = useRef()
-    const passwordRef = useRef()
-    const passwordConfirmRef = useRef()
-    const nameRef = useRef()
-    const surnameRef = useRef()
-    const { signup } = useAuth()
-    const [error, setError] = useState('')
-    const [loading, setLoading] = useState(false)
-    const navigate = useNavigate()
+    const [fullName, setFullname]=useState('');
+    const [email, setEmail]=useState('');
+    const [password, setPassword]=useState('');
 
-    async function handleSubmit(e) {
-        e.preventDefault()
+    const [errorMsg, setErrorMsg]=useState('');
+    const [successMsg, setSuccessMsg]=useState('');
 
-        if (passwordRef.current.value !== passwordConfirmRef.current.value) {
-            return setError('Passwords do not match')    
-        }
-
-        try{
-            setError('')
-            setLoading(true)
-            await signup(emailRef.current.value, passwordRef.current.value)
-            navigate('/')
-        } catch {
-            setError('Failed to create an account. Password length less than 8 characters.')
-        }
-        setLoading(false)
-       
+    const handleSignup=(e)=>{
+        e.preventDefault();
+        // console.log(fullName, email, password);
+        auth.createUserWithEmailAndPassword(email,password).then((credentials)=>{
+            console.log(credentials);
+            fs.collection('users').doc(credentials.user.uid).set({
+                FullName: fullName,
+                Email: email,
+                Password: password
+            }).then(()=>{
+                setSuccessMsg('Signup Successfull. You will now automatically get redirected to Login');
+                setFullname('');
+                setEmail('');
+                setPassword('');
+                setErrorMsg('');
+                setTimeout(()=>{
+                    setSuccessMsg('');
+                    history.push('/login');
+                },3000)
+            }).catch(error=>setErrorMsg(error.message));
+        }).catch((error)=>{
+            setErrorMsg(error.message)
+        })
     }
 
-  return (
-    <>
-      <Card>
-        <Card.Body>
-            <h2 className='text-center mb-4'>Sign Up</h2>
-           
-            {error && <Alert variant='danger'>{error}</Alert>}
-            <Form onSubmit={handleSubmit}>
-              <Form.Group id='name'>
-                    <Form.Label>Name</Form.Label>
-                    <Form.Control type='text' ref={nameRef} required/>
-                </Form.Group>
-                <Form.Group id='surname'>
-                    <Form.Label>Surname</Form.Label>
-                    <Form.Control type='text' ref={surnameRef} required/>
-                </Form.Group>
-                <Form.Group id='email'>
-                    <Form.Label>Email</Form.Label>
-                    <Form.Control type='email' ref={emailRef} required/>
-                </Form.Group>
-                <Form.Group id='password'>
-                    <Form.Label>Password</Form.Label>
-                    <Form.Control type='password' ref={passwordRef} required/>
-                </Form.Group>
-                <Form.Group id='password-confirm'>
-                    <Form.Label>Password Confirmation</Form.Label>
-                    <Form.Control type='password' ref={passwordConfirmRef} required/>
-                </Form.Group>
-                <Button disabled={loading} className='w-100' type='submit'>Sign Up
-                </Button>
-            </Form>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        Already have an account? <Link to='/'>Login In</Link>
-      </div>
-      <div className='w-100 text-center mt-2'>
-        Do you want to be a driver? <Link to='/driversignup'>Sign Up</Link>
-      </div>
-    </>
-  )
+    return (
+        <div className='container'>
+            <br></br>
+            <br></br>
+            <h1>Sign Up</h1>
+            <hr></hr>
+            {successMsg&&<>
+                <div className='success-msg'>{successMsg}</div>
+                <br></br>
+            </>}
+            <form className='form-group' autoComplete="off" onSubmit={handleSignup}>
+                <label>Full Name</label>
+                <input type="text" className='form-control' required
+                onChange={(e)=>setFullname(e.target.value)} value={fullName}></input>
+                <br></br>
+                <label>Email</label>
+                <input type="email" className='form-control' required
+                 onChange={(e)=>setEmail(e.target.value)} value={email}></input>
+                <br></br>
+                <label>Password</label>
+                <input type="password" className='form-control' required
+                 onChange={(e)=>setPassword(e.target.value)} value={password}></input>
+                <br></br>
+                <div className='btn-box'>
+                    <span>Already have an account Login
+                    <Link to="login" className='link'> Here</Link></span>
+                    <button type="submit" className='btn btn-success btn-md'>SIGN UP</button>
+                </div>
+            </form>
+            {errorMsg&&<>
+                <br></br>
+                <div className='error-msg'>{errorMsg}</div>                
+            </>}
+        </div>
+    )
 }
-
