+++ src/Config/Config.js
@@ -4,13 +4,13 @@ import 'firebase/firestore'
 import 'firebase/storage'
 
 const firebaseConfig = {
-  apiKey: "AIzaSyC4kv41W6xnqHliPBGbkTEnWv2DezCzP5w",
-  authDomain: "givealittle-coms3.firebaseapp.com",
-  projectId: "givealittle-coms3",
-  storageBucket: "givealittle-coms3.appspot.com",
-  messagingSenderId: "127398312249",
-  appId: "1:127398312249:web:4a8761a11ba09dc58ad0e8",
-  measurementId: "G-L94XRXSQ40"
+  apiKey: "AIzaSyBs7ggwCZe1kZU-_8tK3hENM8mR4F4nE0o",
+  authDomain: "give-a-little-d525a.firebaseapp.com",
+  projectId: "give-a-little-d525a",
+  storageBucket: "give-a-little-d525a.appspot.com",
+  messagingSenderId: "392348497447",
+  appId: "1:392348497447:web:11be1695e3764c3122c002",
+  measurementId: "G-VTY0MM9XV4"
 };
 
 firebase.initializeApp(firebaseConfig);
