+++ src/index.css
@@ -3,7 +3,7 @@
 *{
   margin: 0;
   padding: 0;
-  font-family: 'Poppins', sans-serif;
+  font-family: 'Montserrat', sans-serif;
 }
 
 .wrapper{
@@ -49,7 +49,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  background-color: #f6f6f6;  
+  background-color:#002752;  
 }
 
 @media(max-width: 450px){
@@ -98,7 +98,7 @@
 
 .navbar .rightside div .navlink{
   text-decoration: none;
-  color: #000;
+  color: #fff;
 }
 
 /* products */
